refactor: update folder structure for clean architecture implementation
