+++ backend/nest-cli.json
@@ -2,6 +2,7 @@
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
   "sourceRoot": "src",
+  "entryFile": "infra/main",
   "compilerOptions": {
     "deleteOutDir": true
   }
