+++ backend/src/infra/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { PassportModule } from '@nestjs/passport'
 import { ClerkStrategy } from './clerk.strategy'
-import { ClerkClientProvider } from 'src/providers/clerk-client.provider'
+import { ClerkClientProvider } from 'src/infra/providers/clerk-client.provider'
 
 @Module({
   imports: [PassportModule, ConfigModule],
