+++ backend/src/infra/auth/clerk.guard.ts
@@ -2,7 +2,7 @@ import { ExecutionContext, Injectable } from '@nestjs/common'
 import { Reflector } from '@nestjs/core'
 import { AuthGuard } from '@nestjs/passport'
 import { Observable } from 'rxjs'
-import { IS_PUBLIC_KEY } from 'src/decorators/public.decorator'
+import { IS_PUBLIC_KEY } from 'src/infra/decorators/public.decorator'
 
 @Injectable()
 export class ClerkAuthGuard extends AuthGuard('clerk') {
