+++ backend/src/infra/auth/clerk.strategy.ts
