+++ backend/src/infra/main.ts
@@ -1,0 +1,8 @@
+import { NestFactory } from '@nestjs/core'
+import { AppModule } from './app.module'
+
+async function bootstrap() {
+  const app = await NestFactory.create(AppModule)
+  await app.listen(process.env.PORT ?? 3000)
+}
+bootstrap()
